Very hacky rewrite of Cookbook Details
